moved data_test.go file contents to data.go file of kubeadm phases
updated code with new line
