+++ cmd/kubeadm/app/cmd/phases/reset/data.go
@@ -39,3 +39,20 @@ type resetData interface {
 	CRISocketPath() string
 	CleanupTmpDir() bool
 }
+
+// a package local type for testing purposes.
+type testData struct{}
+
+// testData must satisfy resetData.
+var _ resetData = &testData{}
+
+func (t *testData) ForceReset() bool                         { return false }
+func (t *testData) InputReader() io.Reader                   { return nil }
+func (t *testData) IgnorePreflightErrors() sets.Set[string]  { return nil }
+func (t *testData) Cfg() *kubeadmapi.InitConfiguration       { return nil }
+func (t *testData) DryRun() bool                             { return false }
+func (t *testData) Client() clientset.Interface              { return nil }
+func (t *testData) CertificatesDir() string                  { return "" }
+func (t *testData) CRISocketPath() string                    { return "" }
+func (t *testData) CleanupTmpDir() bool                      { return false }
+func (t *testData) ResetCfg() *kubeadmapi.ResetConfiguration { return nil }
