+++ cmd/kubeadm/app/cmd/phases/upgrade/data.go
@@ -41,3 +41,22 @@ type Data interface {
 	KubeConfigDir() string
 	KubeletDir() string
 }
+
+// a package local type for testing purposes.
+type testData struct{}
+
+// testData must satisfy Data.
+var _ Data = &testData{}
+
+func (t *testData) EtcdUpgrade() bool                       { return false }
+func (t *testData) RenewCerts() bool                        { return false }
+func (t *testData) DryRun() bool                            { return false }
+func (t *testData) Cfg() *kubeadmapi.UpgradeConfiguration   { return nil }
+func (t *testData) InitCfg() *kubeadmapi.InitConfiguration  { return nil }
+func (t *testData) IsControlPlaneNode() bool                { return false }
+func (t *testData) Client() clientset.Interface             { return nil }
+func (t *testData) IgnorePreflightErrors() sets.Set[string] { return nil }
+func (t *testData) PatchesDir() string                      { return "" }
+func (t *testData) OutputWriter() io.Writer                 { return nil }
+func (t *testData) KubeConfigDir() string                   { return "" }
+func (t *testData) KubeletDir() string                      { return "" }
